rag/requirements.txt
streamlit>=1.28.0
gradio>=4.0.0
plotly>=5.17.0
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
openpyxl>=3.1.0
requests>=2.31.0 